Update Simple Calculator.py
Int is better than float i think

+++ Group-2/Simple Calculator.py
@@ -1,6 +1,6 @@
 # Simple calculator for addition, subtraction, multiplication, and division
-num1 = float(input("Enter the first number: "))
-num2 = float(input("Enter the second number: "))
+num1 = int(input("Enter the first number: "))
+num2 = int(input("Enter the second number: "))
 operation = input("Enter operation (+, -, *, /): ")
 
 if operation == "+":
